+++ src/commands/filterAllTransactionsCommand.ts
@@ -12,7 +12,7 @@
 import { commands, TreeView, window } from "vscode";
 import { CICSCombinedTransactionsTree } from "../trees/CICSCombinedTransactionTree";
 import { CICSTree } from "../trees/CICSTree";
-import { getPatternFromFilter } from "../utils/FilterUtils";
+import { getPatternFromFilter } from "../utils/filterUtils";
 import { PersistentStorage } from "../utils/PersistentStorage";
 
 export function getFilterAllTransactionsCommand(tree: CICSTree, treeview: TreeView<any>) {
