+++ src/commands/filterLocalFileCommand.ts
@@ -12,7 +12,7 @@
 import { commands, ProgressLocation, TreeView, window } from "vscode";
 import { CICSLocalFileTree } from "../trees/CICSLocalFileTree";
 import { CICSTree } from "../trees/CICSTree";
-import { getPatternFromFilter } from "../utils/FilterUtils";
+import { getPatternFromFilter } from "../utils/filterUtils";
 import { PersistentStorage } from "../utils/PersistentStorage";
 
 export function getFilterLocalFilesCommand(tree: CICSTree, treeview: TreeView<any>) {
