+++ src/trees/CICSCombinedTransactionTree.ts
@@ -17,10 +17,10 @@ import { ProfileManagement } from "../utils/profileManagement";
 import { ViewMore } from "./treeItems/utils/ViewMore";
 import { CicsCmciConstants } from "@zowe/cics-for-zowe-cli";
 import { CICSTransactionTreeItem } from "./treeItems/CICSTransactionTreeItem";
-import { toEscapedCriteriaString } from "../utils/toEscapedCriteriaString";
+import { toEscapedCriteriaString } from "../utils/FilterUtils";
 import { CICSRegionsContainer } from "./CICSRegionsContainer";
 import { TextTreeItem } from "./treeItems/utils/TextTreeItem";
-import { getIconPathInResources } from "../utils/getIconPath";
+import { getIconPathInResources } from "../utils/profileUtils";
 
 export class CICSCombinedTransactionsTree extends TreeItem {
   children: (CICSTransactionTreeItem | ViewMore) [] | [TextTreeItem] | null;
